Community cycle management
- Add new cycle management dashboard

#2835
