+++ api/admin/service.js
@@ -323,6 +323,9 @@ async function getCommunity (id) {
   community.tasks = (await dao.Task.db.query(dao.query.communityTaskStateQuery, id)).rows[0];
   community.tasks.totalCreated = Object.values(community.tasks).reduce((a, b) => { return a + parseInt(b); }, 0);
   community.users = (await dao.User.db.query(dao.query.communityUsersQuery, id)).rows[0];
+  if (community.duration == 'Cyclical') {
+    community.cycles = await dao.Cycle.find('community_id = ?', community.communityId);
+  }
   return community;
 }
